Move FileOpenRd/FileOpenWr/FileClose to Env class
Delete lotus/core/graph/utils.h
Add a new method "GetSelfPid" to Env class. Mainly for setting random seed.
Add "-Wno-error=invalid-partial-specialization" to clang cflags, because we reverted eigen to an older version and there is a warning in eigen that we can't fix.
Remove 'virtual' keyword from ILotusOpSchemaCollection::GetSchema because this function is already declared 'final'
Add a new macro "LOTUS_MUST_USE_RESULT", and apply it to GetAttr() functions.

Related work items: #481

+++ lotus/core/framework/bfc_arena.cc
@@ -46,7 +46,7 @@ BFCArena::~BFCArena() {
 }
 
 BFCArena::Chunk* BFCArena::ChunkFromHandle(ChunkHandle h) {
-  LOTUS_ENFORCE(h < static_cast<int>(chunks_.size()));
+  LOTUS_ENFORCE(h < chunks_.size());
   return &(chunks_[h]);
 }
 
@@ -81,7 +81,7 @@ bool BFCArena::Extend(size_t rounded_bytes) {
 
     // Try allocating less memory.
     while (mem_addr == nullptr) {
-      bytes = RoundedBytes((size_t)(bytes * kBackpedalFactor));
+      bytes = RoundedBytes(static_cast<size_t>(bytes * kBackpedalFactor));
       if (bytes < rounded_bytes) break;
       mem_addr = device_allocator_->Alloc(bytes);
     }
